Day 5 - Part 1 - Example
After downloading the input, I realize how memory inefficient this is.
The main example may run — in which case, I wouldn't change this — but
most likely, it may break because of the too much memory. I would like
to remember this though, as a translation of problem statement to code