Day 8 - Part 1
Gotta  say, it's awfully quiet coding without Copilot. I am solving this
on a plane. There's no internet. So Copilot is asleep. I miss that lad.

That said, this version is yet to be tested with the original input. I
forgot to download the input before getting on the plane. Subsequent
days done on the plane have their inputs downloaded. But for oversight
reasons, since Day 8's question page was already open in a tab, I did
not get the input for the actual question.

Now I write this... I realize I would not be able to verify if my
solution is correct with the Advent anyways. No internet. What these
exercises would do, asides curing the boredom of flying economy, is to
verify the solution with the example cases.