Day 10 - Part 0
Had to log this part because it's a step in getting the corruption score
of the instructions. Solving this first part for me was in two steps.

  1. Figure out which of the strings are corrupt, and
  2. Calculate the cost of the corruption

Gotta love the terminology now, don't we?

This zeroth part calculates returns the number of corrupt instructions
in the set. From the example given, there are 5 corrupt instructions;
this is shown in the question. Looks like there are 49 corrupt
instructions from the actual input. That's half the size of the input!